added discovery struct and node setup
I added a structure that may be useful for our discovery message, I added the functions for setting its attributes (none for getting yet). I also added functionality for establishing the nodes plugin and opening the node, and passing control to the physical module. Comments included on each of these points to further explain what is being performed.